Fix: Clean up code and reorder imports
This commit addresses minor code cleanup and import reordering.

Key changes:
- **SalesMapper.kt:**
    - Reordered import statements for consistency.
- **RemoteKey.kt:**
    - Removed an unnecessary blank line at the end of the data class definition.

+++ app/src/main/java/com/blipblipcode/distribuidoraayl/core/local/entities/reportSale/SaleDataEntity.kt
@@ -11,7 +11,7 @@ data class SaleDataEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long = 0L,
     val uid: String? = null,
-    val number: Int,
+    val number: Long,
     val clientRut: String,
     val date: Long,
     val token: String?,
